Setup index.js and index.css
The index.js will contain the
homeModule when the page
first loads.

Create a css file and import it
in the index.js. Select the
and save the references of
these elements on respective
constants. These buttons will
be used to render their
respective modules.

Create a function that creates
a div, gives it a class and a
textContent and append it to the
the #content div (container).

Change the text color of
the #content on the index.css to
confirm if the css is being
imported successfully.