Setup the menu module
Create a menu directory and
create the menu module js file
and the menu css files inside
it. This directory contains
everything about the menu module.

In the menu module js file, first
import the menu CSS.
Export a function that selects
the #container, create a div,
give it a class and a text content
and append it to #container.

Add a style in the menu css to
confirm if the css is being
imported successfully.

Import this function into
index.js. Attach a click
eventListener to the menuButton
so that when it is clicked,
the textContent of container
is cleared before the function
tha was exported by the menu
module is called. This effectively
'wipes' out the homeContainer
contents and renders the
menuContainer contents.