Setup the about module
Set the about module just
as the menu module was set
in the previous commit.
Create the respective
directory and the js and
css files.
Export the respective function
and import it in the index.js.

In index.js, attach an event
listener to the aboutButton
and have it clear the textContent
of container and call the
just imported function
(createAboutModule).