Start styling
Add a div in the index.html
that would serve as the footer.

In the index.css, add basic resets
and styling. Give the body a grid
display and grid template rows
of auto 1fr auto so that the nav
and .footer will be automatically
sized to fit their contents while
the 1fr  will distribute the
remaining space to the #content.

Style the nav with a flex display
and justify its contents center.
Give a common padding to the nav
and the .footer.