Create home module content
Create the structure of the
hero in such a way that a
heroLeft div contains the
heroText div and the heroButton.
The heroText div contains the
heroH1 and heroParagraph. The
heroRight div will contain an
image.

Add the respective classes and
text content to the respective
elements. Use template literals
for the heroParagraph text content
so that it can span multiple
lines for better readability
on the text editor and
still be treated as a single
string.