Start styling home container
Style the home container with
a flex display, justify its
contents center and give it
a flex-wrap property of wrap.
The wrap property will make
the hero image container to
break to a new line when the
home container width is smaller
than the combined flex-basis
(400px + 400px) of the
 hero text container and the
hero image container - on
smaller screens

Set the flex-basis of both
hero text container and the
hero image container.

Set the border box property
on the * selector and also
give the body a gap.