Import and add hero image
Configure the webpack config
so that webpack can use the
built in asset modules to
handle importing images.

Create a home-assets folder
and an images folder inside it.
Add the image file in the images
folder and import it in the
index.js file. Create a new
HTMLImageElement object, give
it a relevant class and set
its src attribute to the
imported image.

Append this new HTMLImageElement
object (image) to heroRight.

In the index.css, reduce the
body's gap so that the content
will fit into the
page without a vertical scroll
bar (On large screens).
Give the hero-image a 100%
width so that it will occupy
the full width of its container.