Style buttons and hero text container
Increase the font-size of button,
add padding and border radius.

Style the hero text container
with a grid display and a gap.

Style the hero button with a
margin top value to give it
some space between it and the
hero text container.

Remove the gap from body because
the new styles applied to
especially button increased the
page content. The gap in body
will cause a vertical scroll.