Style the menu items container
Style it with a grid display.
Style it with template columns
such that it repeats a pattern
of columns where the number of
columns is determined by the
space in the container. The
minmax defines the minimum
and maximum widths respectively
of each column.

Style it with a gap and padding
left and right values.

Give the menu items images a
100% width.

Add a media query that reduces the
padding left and right values
on smaller screens else the page
will have a horizontal scrollbar
(on smaller screens).