Style the menu container
Style it with a grid display
and a gap. Add a padding top
value so that there is some
space between the menu h3
and the nav.

Style the menu h3 with a
justify-self value of center
so that it is justified to
the center.

Increase the padding left and
right on the menu items
container.