Create divs for overlay
The menu images will have
an overlay that will give
a brief description of the
image.

Create overlay divs with
respect to each menu item and
set the textContent of these
divs respectively. Append these
newly created divs to their
respective menu items (as a
sibling to the menu item images).

On line 69, add all the menu items
to an array, loop through the
array and add a class to each
one of the menu items.
This is to make it easier to
select them and their children
in the CSS and style
 accordingly.