Style the overlay with CSS
Style the menu item with
a relative position so that
the overlay (menu-item > div)
can be positioned with
absolute. Give it an ease
transition of 0.5s so that
Style its hover pseudo class
to transform by 1.04 scale.
The transform property will
render according to the
transition mentioned above.

Style the menu-item > div with
an absolute position and give
it a bottom value of 0 so that
it will be at the bottom
of the menu item.
Set its background to a
transparent black color and
its (text) color to a shade
of white. Increase its
font size, add padding and
align its text to the center.
Also style the border bottom
left and right radiuses.

Change the selector for
the menu items images from
from general descendant
combinator
(.menu-items-container img)
to direct descendant
combinator (.menu-item > img)