Populate the about module
Create an about story div
and the contact container
div. Create the google
maps container and the
address container. Set the
innerHTML of the google maps
container to an iframe of
an embedded google map. Create
the address, city and
phone divs.

Append the about story div
and contact container div
to the about container.
Append the google maps container
and the address container to
to the contact container.
Append the address, city and
phone divs to address container.