Style the about container
Give the about container
a grid display and a gap.

In about.module.js, remove
the inline width and height
style from the map iframe
so that CSS can be used to
style better.

Style contact container with
a flex display and a flex wrap
value of wrap. Justify its
contents center and give it a
gap.

Set the flex basis of the map
and address containers so that
they can have a initial width
from which they will grow or
shrink as needed.

Set the width and height of
the iframe - width so that
it occupies 100% of its
container and height because
if it is not set, it is not
long enough.