Style the about story
The about story div was
occupying the full width
of its container. Style it
such that it takes 50%
of its container and give
it a justify self property
of center.

Add a media query that makes
the about story take up 80%
of its container width on
smaller devices. In the media
query, set a lower height
value for the iframe and the
address container.