Add a nav title and style
In the index.html, add a
h1 nav-title. Wrap the
nav items in a div so they
can be styled better

In index.css, remove the
gap from nav and justify
its content with
space-around, to put
space around the nav title
and the nav items.

Align the nav title to the
center so that it aligns
horizontally with the
nav items.

Add a media query that
changes the nav display to
grid, places its content
center and gives it a gap.
This makes the nav title
to be on top the nav items
on smaller devices.

In the media query, justify
the nav title center so it
aligns vertically with the
nav items. Also reduce the
buttons font size and padding.