Add styles to body
In the index.css, add a gap
to the body so
there would be space between
its items (header, content and
footer). Give body a color,
a background color and a
font-family with fallback
fonts.

Increase nav and footer's
paddings.to remove the small
whitespace after footer. Also
give them a background color.

Increase the font size of
hero text container and
give container a
background color.

In the media query, reduce
the value of the gap of body
so that the items of body
do not look too far away from
each other on smaller screens

In the index.js, update the
hero created element from a
h1 to a h2 and change its
class name. Also change its
variable name and update
all instances of the name
to reflect the new name.