Style nav items
In index.css,
Style nav items with a
flex display and a gap.

Give all buttons a border
radius and edit the previous
button selector to select
the nav buttons. Reduce the
font size of the nav buttons
and give them a background
color and a color.

Increase the hero button's
padding and its font size.
Give it a bold font-weight,
a color and a background
color.

In index.js, update the
hero h2's text content.