Add styles to the about.css
Increase the size of the
about story. Add a border
and a border radius to
map container, the iframe
and address container. If
a border radius is not
added to the iframe, its
edges will stick out of
the map container

Reduce the height of the
iframe so that the page
would not overflow to have
a vertical scroll bar.

STyle the address container
with a flex display, change
its direction to column and
a gap. Increase its
font size and its padding.