Add more styles to index and menu
In index.css, add a border radius
to the hero image.

In the media
query, decrease the padding and
font size of the hero button.
Also style the hero left container
with a grid display so that the
hero button can span the width
of its container. Align its text
to the center. This looks better
on smaller screens.

In menu.css, reduce the font size
of the menu items overlay divs.
Style the menu item images with
a border.