Update footer and style
In the index.html, update the
footer text. Also update the
favicon to an emoji.

In index.css, align the footer
text to the center.

Style the nav items buttons
hover pseudo class with a
background color and a text
color.

Style the hero button
hover pseudo class with a
background color.

Increase the gap for body
to create some space and have
the footer cover some white
space at the bottom of the
page.

In the about module js, update
the street text content.